refactor: img upload to utils
