+++ src/components/Utils/index.js
@@ -1,1 +1,3 @@
 export { default as dateToText } from './dateToText';
+export { default as imgSrcToFile } from './imgSrcToFile';
+export { default as loadImg } from './loadImg';
